Initial commit using pmapper for other escalations
This is an attempt to use pmapper with other escalations in
iam__priv_esc. It attempts to define a common interface for
these, it run's checks, finds a list of paths and asks you
which path to run.

This is using some local forks of other projects at the moment
so unless those are set up this won't work for you right now.