Use subclassing for tracking edges
Haven't tested this yet, but thinking this general concept makes
sense.

Instead of using seperate checkers we have one root class which
will be the only one attached to checker_map. We track checkers
via cls.__subclasses__ which can in turn either be a proxy to
a larger group of checkers or be a checker it self.

Building the graph is already pretty slow, checkers run at O(n),
while node size is O(n^2) or around there. We'll be adding more
checker's with this approach, but we can reduce the run time
by doing a simple precheck's before we make source/dest pairs
out of the nodes. This is handled through filter_source and
filter_dest methods.

At least in theory.. maybe